+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 *   (improvement) Return `bool` in `SortableHandler::sortElementBefore()` instead of throwing an exception.
 *   (deprecation) Deprecate `InvalidSortOperationException`, as it is not used anymore.
+*   (improvement) Mark all sortable handlers as `final`.